+++ l10n/ms_MY.js
@@ -1,0 +1,16 @@
+OC.L10N.register(
+    "richdocuments",
+    {
+    "Description" : "Keterangan",
+    "Error" : "Ralat",
+    "Nickname" : "Nama Samaran",
+    "Cancel" : "Batal",
+    "Save" : "Simpan",
+    "Remove from favorites" : "Remove from favorites",
+    "Add to favorites" : "Add to favorites",
+    "Download" : "Muat turun",
+    "Create" : "Buat",
+    "Password" : "Kata laluan",
+    "OK" : "OK"
+},
+"nplurals=1; plural=0;");