+++ l10n/ug.js
@@ -1,0 +1,16 @@
+OC.L10N.register(
+    "richdocuments",
+    {
+    "Description" : "چۈشەندۈرۈش",
+    "Error" : "خاتالىق",
+    "Nickname" : "تەخەللۇس",
+    "Cancel" : "ۋاز كەچ",
+    "Save" : "ساقلا",
+    "Remove from favorites" : "Remove from favorites",
+    "Add to favorites" : "Add to favorites",
+    "Download" : "چۈشۈر",
+    "Create" : "قۇر",
+    "Password" : "ئىم",
+    "OK" : "جەزملە"
+},
+"nplurals=1; plural=0;");