+++ lib/Controller/WopiController.php
@@ -548,8 +548,9 @@ class WopiController extends Controller {
 	public function putFile($fileId,
 							$access_token) {
 		list($fileId, ,) = Helper::parseFileId($fileId);
-		$isPutRelative = ($this->request->getHeader('X-WOPI-Override') === 'PUT_RELATIVE');
-		$isRenameFile = ($this->request->getHeader('X-WOPI-Override') === 'RENAME_FILE');
+		$override = $this->request->getHeader('X-WOPI-Override');
+		$isPut = ($override === 'PUT');
+		$result = new JSONResponse([], Http::STATUS_OK);
 
 		$wopi = $this->wopiMapper->getWopiForToken($access_token);
 		if (empty($wopi) || !$wopi->getCanwrite()) {
@@ -559,55 +560,49 @@ class WopiController extends Controller {
 		if (empty($file)) {
 			return new JSONResponse([], Http::STATUS_FORBIDDEN);
 		}
-		try {
-			if ($isPutRelative) {
-				$suggested = $this->request->getHeader('X-WOPI-SuggestedTarget');
-				$suggested = iconv('utf-7', 'utf-8', $suggested);
 
-				if ($suggested[0] === '.') {
-					$path = dirname($file->getPath()) . '/New File' . $suggested;
-				}
-				else if ($suggested[0] !== '/') {
-					$path = dirname($file->getPath()) . '/' . $suggested;
-				}
-				else {
-					$userFolder = $this->rootFolder->getUserFolder($wopi->getEditorUid());
-					$path = $userFolder->getPath() . $suggested;
-				}
+		// Set the user to register the change under his name
+		$this->userScopeService->setUserScope($wopi->getEditorUid());
+		$this->userScopeService->setFilesystemScope($isPutRelative ? $wopi->getEditorUid() : $wopi->getUserForFileAccess());
 
-				if ($path === '') {
-					return new JSONResponse([
-						'status' => 'error',
-						'message' => 'Cannot create the file'
-					]);
+		$lck = $this->request->getHeader('X-WOPI-Lock');
+		$fLock = $this->lockMapper->find($fileId);
+		if ($isPut) {
+			if (!empty($fLock))
+			{
+				if ($fLock->getValue() !== $lck)
+				{
+					$result->setStatus(Http::STATUS_CONFLICT);
+					$result->addHeader('X-WOPI-Lock', $fLock->getValue());
 				}
-
-				// create the folder first
-				if (!$this->rootFolder->nodeExists(dirname($path))) {
-					$this->rootFolder->newFolder(dirname($path));
-				}
-
-				// create a unique new file
-				$path = $this->rootFolder->getNonExistingName($path);
-				$this->rootFolder->newFile($path);
-				$file = $this->rootFolder->get($path);
-			} else {
-				$wopiHeaderTime = $this->request->getHeader('X-LOOL-WOPI-Timestamp');
-
-				if (!empty($wopiHeaderTime) && $wopiHeaderTime !== Helper::toISO8601($file->getMTime() ?? 0)) {
-					$this->logger->debug('Document timestamp mismatch ! WOPI client says mtime {headerTime} but storage says {storageTime}', [
-						'headerTime' => $wopiHeaderTime,
-						'storageTime' => Helper::toISO8601($file->getMTime() ?? 0)
-					]);
-					// Tell WOPI client about this conflict.
-					return new JSONResponse(['LOOLStatusCode' => self::LOOL_STATUS_DOC_CHANGED], Http::STATUS_CONFLICT);
+			}
+			else
+			{
+				if ($file->getSize() > 0)
+				{
+					$result->setStatus(Http::STATUS_CONFLICT);
+					$result->addHeader('X-WOPI-Lock', '');
 				}
 			}
 
+			$wopiHeaderTime = $this->request->getHeader('X-LOOL-WOPI-Timestamp');
+			if (!empty($wopiHeaderTime) && $wopiHeaderTime !== Helper::toISO8601($file->getMTime() ?? 0)) {
+				$this->logger->debug('Document timestamp mismatch ! WOPI client says mtime {headerTime} but storage says {storageTime}', [
+					'headerTime' => $wopiHeaderTime,
+					'storageTime' => Helper::toISO8601($file->getMTime() ?? 0)
+				]);
+				// Tell WOPI client about this conflict.
+				return new JSONResponse(['LOOLStatusCode' => self::LOOL_STATUS_DOC_CHANGED], Http::STATUS_CONFLICT);
+			}
+		}
+		else {
+			$result->setStatus(Http::STATUS_NOT_IMPLEMENTED);
+		}
+		if ($result->getStatus() !== Http::STATUS_OK)
+			return $result;
+		try {
 			$content = fopen('php://input', 'rb');
 
-			// Set the user to register the change under his name
-			$this->userScopeService->setUserScope($wopi->getEditorUid());
 			$this->lockHooks->setLockBypass(true);
 			try {
 				$this->retryOperation(function () use ($file, $content){
@@ -618,17 +613,10 @@ class WopiController extends Controller {
 				return new JSONResponse(['message' => 'File locked'], Http::STATUS_INTERNAL_SERVER_ERROR);
 			}
 
-			if ($isPutRelative) {
-				// generate a token for the new file (the user still has to be
-				// logged in)
-				list(, $wopiToken) = $this->tokenManager->getToken($file->getId(), null, $wopi->getEditorUid());
-
-				$wopi = 'index.php/apps/wopi/wopi/files/' . $file->getId() . '_' . $this->config->getSystemValue('instanceid') . '?access_token=' . $wopiToken;
-				$url = $this->urlGenerator->getAbsoluteURL($wopi);
-
-				return new JSONResponse([ 'Name' => $file->getName(), 'Url' => $url ], Http::STATUS_OK);
+			if ($wopi->hasTemplateId()) {
+				$wopi->setTemplateId(null);
+				$this->wopiMapper->update($wopi);
 			}
-
 			return new JSONResponse(['LastModifiedTime' => Helper::toISO8601($file->getMTime())]);
 		} catch (\Exception $e) {
 			$this->logger->logException($e, ['level' => ILogger::ERROR,	'app' => 'wopi', 'message' => 'getFile failed']);
@@ -641,6 +629,8 @@ class WopiController extends Controller {
 	 * Expects a valid token in access_token parameter.
 	 * Just actually routes to the PutFile, the implementation of PutFile
 	 * handles both saving and saving as.* Given an access token and a fileId, replaces the files with the request body.
+	 *
+	 * FIXME Cleanup this code as is a lot of shared logic between putFile and putRelativeFile
 	 *
 	 * @PublicPage
 	 * @NoCSRFRequired
@@ -688,7 +678,7 @@ class WopiController extends Controller {
 					{
 						$result = new JSONResponse([], Http::STATUS_CONFLICT);
 						$result->addHeader('X-WOPI-Lock', $fLock->getValue());
-						return result;
+						return $result;
 					}
 					$suggested = $this->request->getHeader('X-WOPI-RequestedName');
 
@@ -712,6 +702,7 @@ class WopiController extends Controller {
 					// create a unique new file
 					$path = $this->rootFolder->getNonExistingName($path);
 					$file = $file->move($path);
+					return new JSONResponse([], Http::STATUS_OK);
 				} else {
 					$suggested = $this->request->getHeader('X-WOPI-SuggestedTarget');
 					$relative = $this->request->getHeader('X-WOPI-RelativeTarget');
@@ -770,9 +761,9 @@ class WopiController extends Controller {
 			}
 
 			$content = fopen('php://input', 'rb');
-
 			// Set the user to register the change under his name
 			$this->userScopeService->setUserScope($wopi->getEditorUid());
+			$this->userScopeService->setFilesystemScope($wopi->getEditorUid());
 			$this->lockHooks->setLockBypass(true);
 			try {
 				$this->retryOperation(function () use ($file, $content){
@@ -788,8 +779,8 @@ class WopiController extends Controller {
 
 			$wopi = 'index.php/apps/wopi/wopi/files/' . $file->getId() . '_' . $this->config->getSystemValue('instanceid') . '?access_token=' . $wopiToken;
 			$url = $this->urlGenerator->getAbsoluteURL($wopi);
-
-			return new JSONResponse([ 'Name' => $file->getName(), 'Url' => $url ], Http::STATUS_OK);
+			$editUrl = $this->urlGenerator->getAbsoluteURL('index.php/apps/richdocuments/edit/' . $file->getId() . '_' . $this->config->getSystemValue('instanceid'));
+			return new JSONResponse([ 'Name' => $file->getName(), 'Url' => $url, 'HostEditUrl' => $editUrl, 'HostViewUrl' => $editUrl ], Http::STATUS_OK);
 		} catch (\Exception $e) {
 			$this->logger->logException($e, ['level' => ILogger::ERROR,	'app' => 'wopi', 'message' => 'putRelativeFile failed']);
 			return new JSONResponse([], Http::STATUS_INTERNAL_SERVER_ERROR);
@@ -843,14 +834,7 @@ class WopiController extends Controller {
 		} else {
 			// Unless the editor is empty (public link) we modify the files as the current editor
 			// TODO: add related share token to the wopi table so we can obtain the
-			$editor = $wopi->getEditorUid();
-
-			// Use the actual file owner no editor is available
-			if ($editor === null) {
-				$editor = $wopi->getOwnerUid();
-			}
-
-			$userFolder = $this->rootFolder->getUserFolder($editor);
+			$userFolder = $this->rootFolder->getUserFolder($wopi->getUserForFileAccess());
 			$files = $userFolder->getById($wopi->getFileid());
 			if (isset($files[0]) && $files[0] instanceof File) {
 				$file = $files[0];