+++ templates/admin.php
@@ -5,83 +5,23 @@ script('files', 'jquery.fileupload');
 
 /** @var array $_ */
 ?>
-<div class="section" id="wopi">
-	<h2>
-		<?php p($l->t('Office Online')) ?>
-	</h2>
-	<p>
-		<label for="wopi_url"><?php p($l->t('URL (and Port) of Office Online-server')) ?></label>
-		<br />
-		<input type="text" name="wopi_url" id="wopi_url" placeholder="https://office.example.org:9980" value="<?php p($_['wopi_url'])?>"><button type="button" id="wopi_apply"><?php p($l->t('Apply')) ?></button>
-		<span id="documents-admin-msg" class="msg"></span>
-		<br />
-		<em><?php p($l->t('Office Online requires a seperate server acting as a WOPI-like Client to provide editing capabilities.')) ?></em>
-	</p>
-	<p>
-		<input type="checkbox" class="checkbox" id="disable_certificate_verification" <?php p($_['disable_certificate_verification'] === 'yes' ? 'checked' : '') ?> />
-		<label for="disable_certificate_verification"><?php p($l->t('Disable certificate verification (insecure)')) ?></label><br />
-		<em><?php p($l->t('Enable if your Office Online server uses a self signed certificate')) ?></em>
-	</p>
-	<p>
-		<input type="checkbox" class="use-groups-enable checkbox" id="use_groups_enable-wopi" />
-		<label for="use_groups_enable-wopi"><?php p($l->t('Restrict usage to specific groups')) ?></label>
-		<br/>
-		<em><?php p($l->t('Office Online is enabled for all users by default. When this setting is active, only members of the specified groups can use it.')) ?></em><br />
-		<input type="hidden" id="use_group_select" value="<?php p($_['use_groups'])?>" title="<?php p($l->t('All')); ?>">
-	</p>
-	<p>
-		<input type="checkbox" class="edit-groups-enable checkbox" id="edit_groups_enable-wopi" />
-		<label for="edit_groups_enable-wopi"><?php p($l->t('Restrict edit to specific groups')) ?></label>
-		<br/>
-		<em><?php p($l->t('All users can edit documents with Office Online by default. When this setting is active, only the members of the specified groups can edit and the others can only view documents.')) ?></em><br />
-		<input type="hidden" id="edit_group_select" value="<?php p($_['edit_groups'])?>" title="<?php p($l->t('All')); ?>">
-	</p>
+<div id="admin-vue" data-initial="<?php p(json_encode($_['settings'], true)); ?>"></div>
 
-	<h3>
-		<?php p($l->t('Advanced Settings')) ?>
-	</h3>
-	<p>
-		<input type="checkbox" class="doc-format-ooxml checkbox" id="doc_format_ooxml_enable-wopi" <?php p($_['doc_format'] === 'ooxml' ? 'checked' : '') ?> />
-		<label for="doc_format_ooxml_enable-wopi"><?php p($l->t('Use Office Open XML (OOXML) instead of OpenDocument Format (ODF) by default for new files')) ?></label>
-	</p>
-	<p>
-	<input type="checkbox" class="checkbox" id="enable_external_apps_cb-wopi" <?php p($_['external_apps'] !== '' ? 'checked' : '') ?> />
-	<label for="enable_external_apps_cb-wopi"><?php p($l->t('Enable access for external apps')) ?></label>
-	<div id="enable-external-apps-section" class="indent <?php if ($_['external_apps'] === '') p('hidden') ?>" >
-		<div id="external-apps-section">
-			<input type="hidden" id="external-apps-raw" name="external-apps-raw" value="<?php p($_['external_apps']) ?>">
-		</div>
-
-		<button type="button" id="external-apps-save-button"><?php p($l->t('Save')) ?></button>
-		<button type="button" id="external-apps-add-button"><?php p($l->t('Add')) ?></button>
-		<span id="enable-external-apps-section-msg" class="msg"></span>
-	</div>
-	</p>
-	<p>
-	<input type="checkbox" class="checkbox" id="enable_canonical_webroot_cb-wopi" <?php p($_['canonical_webroot'] !== '' ? 'checked' : '') ?> />
-	<label for="enable_canonical_webroot_cb-wopi"><?php p($l->t('Use Canonical webroot')) ?></label>
-	<div id="enable-canonical-webroot-section" class="indent <?php if ($_['canonical_webroot'] === '') p('hidden') ?>" >
-		<input type="text" id="canonical-webroot" name="canonical-webroot-name" value="<?php p($_['canonical_webroot']) ?>">
-		<br/>
-		<p class="rd-settings-documentation"><em><?php p($l->t('Canonical webroot, in case there are multiple, for Office to use. Provide the one with least restrictions. Eg: Use non-shibbolized webroot if this instance is accessed by both shibbolized and non-shibbolized webroots. You can ignore this setting if only one webroot is used to access this instance.')) ?></em></p>
-	</div>
-	</p>
-</div>
-<?php if ($_['templatesAvailable'] === true) { ?>
-<form class="section" id="wopi-templates" method="post" action="/template/">
+<?php if ($_['settings']['templatesAvailable'] === true) { ?>
+<form class="section" id="richdocuments-templates" method="post" action="/template/">
 	<input class="hidden-visually" id="add-template" type="file" />
 	<h2>
 		<?php p($l->t('Global templates')) ?>
 		<label for="add-template" class="icon-add" title="<?php p($l->t('Add a new template')); ?>"></label>
 	</h2>
-	<div id="emptycontent" class="<?php p(empty($_['templates'])?:'hidden') ?>">
+	<div id="emptycontent" class="<?php p(empty($_['settings']['templates'])?:'hidden') ?>">
 		<div class="icon-file"></div>
 		<h2>
 			<?php p($l->t('No templates defined.')); ?>
 		</h2>
 		<label for="add-template"><?php p($l->t('Add a new one?')); ?></label>
 	</div>
-	<ul class="<?php p(!empty($_['templates'])?:'hidden') ?>">
+	<ul class="<?php p(!empty($_['settings']['templates'])?:'hidden') ?>">
 		<li class="hidden template-model">
 			<figure>
 				<img src="" alt="<?php p($l->t('template preview')) ?>" />
@@ -90,7 +30,7 @@ script('files', 'jquery.fileupload');
 			<a href="" class="delete-template icon-delete"></a>
 			<div class="delete-cover"></div>
 		</li>
-		<?php foreach ($_['templates'] as $template) {?>
+		<?php foreach ($_['settings']['templates'] as $template) {?>
 			<li>
 				<figure>
 					<?php if (isset($template['preview'])) { ?>
@@ -108,4 +48,3 @@ script('files', 'jquery.fileupload');
 </form>
 <?php } ?>
 
-<div id="admin-vue" data-initial="<?php p(json_encode($_['settings'], true)); ?>"></div>